+++ bin/just-bin-it.py
@@ -4,42 +4,14 @@ import logging
 import os
 import sys
 import time
-import graphyte
 
 sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.realpath(__file__))))
 from just_bin_it.endpoints.config_listener import ConfigListener
 from just_bin_it.endpoints.kafka_consumer import Consumer
-from just_bin_it.endpoints.kafka_producer import Producer
 from just_bin_it.endpoints.kafka_tools import are_kafka_settings_valid
-from just_bin_it.endpoints.sources import EventSource, SimulatedEventSource
-from just_bin_it.histograms.histogrammer import create_histogrammer, parse_config
-
-
-class StatisticsPublisher:
-    def __init__(self, server, port, prefix, metric):
-        """
-        Constructor.
-
-        :param server: The Graphite server to send to.
-        :param port: The Graphite port to send to.
-        :param prefix: The data prefix in Graphite.
-        :param metric: The name to serve data in Graphite as.
-        """
-        self.server = server
-        self.port = port
-        self.prefix = prefix
-        self.metric = metric
-
-        # Initialise the connection
-        graphyte.init(server, port=port, prefix=prefix)
-
-    def send_histogram_stats(self, hist_stats):
-        for i, stat in enumerate(hist_stats):
-            graphyte.send(
-                f"{self.metric}{i}",
-                stat["sum"],
-                timestamp=stat["last_pulse_time"] / 10 ** 9,
-            )
+from just_bin_it.histograms.histogrammer import parse_config
+from just_bin_it.histograms.histogram_process import HistogramProcess
+from just_bin_it.utilities.statistics_publisher import StatisticsPublisher
 
 
 def load_json_config_file(file):
@@ -63,7 +35,6 @@ class Main:
         self,
         config_brokers,
         config_topic,
-        one_shot,
         simulation,
         initial_config=None,
         stats_publisher=None,
@@ -73,14 +44,12 @@ class Main:
 
         :param config_brokers: The brokers to listen for the configuration commands on.
         :param config_topic: The topic to listen for commands on.
-        :param one_shot: Histogram some data then plot it. Does not publish results.
         :param simulation: Run in simulation mode.
         :param initial_config: A histogram configuration to start with.
         :param stats_publisher: Publisher for the histograms statistics.
         """
         self.event_source = None
         self.histogrammer = None
-        self.one_shot = one_shot
         self.simulation = simulation
         self.initial_config = initial_config
         self.config_brokers = config_brokers
@@ -108,26 +77,8 @@ class Main:
 
                 try:
                     logging.warning("New configuration command received")
-
-                    # TODO: stop existing processes here?
-
-                    start, stop, hist_configs = parse_config(msg)
-
-                    # TODO: Check kafka settings etc here?
-
-                    for hist in hist_configs:
-                        from just_bin_it.histograms.histogram_process import (
-                            HistogramProcess,
-                        )
-
-                        p = HistogramProcess(hist, start, stop)
-                        self.hist_process.append(p)
-
-                    # TODO: Handle different msg types
-                    # self.handle_command_message(msg)
-                    # if not self.one_shot:
-                    #     # Publish initial empty histograms.
-                    #     self.histogrammer.publish_histograms()
+                    logging.warning("%s", msg)
+                    self.handle_command_message(msg)
                 except Exception as error:
                     logging.error("Could not handle configuration: %s", error)
 
@@ -156,51 +107,28 @@ class Main:
             Consumer(self.config_brokers, [self.config_topic])
         )
 
-    def configure_histograms(self, config):
-        """
-        Configure histogramming based on the supplied configuration.
-
-        :param config: The configuration.
-        """
-        producer = Producer(config["data_brokers"])
-        self.histogrammer = create_histogrammer(
-            producer, config, int(time.time() * 1000)
-        )
-        if self.histogrammer.start:
-            self.event_source.seek_to_time(self.histogrammer.start)
-
-    def configure_event_source(self, config):
-        """
-        Configure the event source.
-
-        :param config: The configuration.
-        :return: The new event source.
-        """
-        # Check brokers and data topics exist
-        if not are_kafka_settings_valid(config["data_brokers"], config["data_topics"]):
-            raise Exception("Invalid event source settings")
-
-        consumer = Consumer(config["data_brokers"], config["data_topics"])
-        event_source = EventSource(consumer)
-
-        return event_source
-
     def handle_command_message(self, message):
         """
         Handle the message received.
 
         :param message: The message.
         """
-
         if message["cmd"] == "restart":
+            # TODO: Sort this out
             self.histogrammer.clear_histograms()
         elif message["cmd"] == "config":
-            if self.simulation:
-                logging.warning("RUNNING IN SIMULATION MODE")
-                self.event_source = SimulatedEventSource(message)
-            else:
-                self.event_source = self.configure_event_source(message)
-            self.configure_histograms(message)
+            logging.info("Stopping existing processes")
+            for proc in self.hist_process:
+                proc.stop_process()
+            self.hist_process.clear()
+
+            start, stop, hist_configs = parse_config(message)
+
+            # TODO: Check kafka settings etc here?
+
+            for hist in hist_configs:
+                p = HistogramProcess(hist, start, stop)
+                self.hist_process.append(p)
         else:
             raise Exception(f"Unknown command type '{message['cmd']}'")
 
@@ -234,14 +162,6 @@ if __name__ == "__main__":
         "--graphite-config-file",
         type=str,
         help="configuration file for publishing to Graphite",
-    )
-
-    parser.add_argument(
-        "-o",
-        "--one-shot-plot",
-        action="store_true",
-        help="runs the program until it gets some data, plot it and then exit."
-        " Used for testing",
     )
 
     parser.add_argument(
@@ -283,11 +203,6 @@ if __name__ == "__main__":
         logging.basicConfig(format="%(asctime)s - %(message)s", level=logging.INFO)
 
     main = Main(
-        args.brokers,
-        args.topic,
-        args.one_shot_plot,
-        args.simulation_mode,
-        init_hist_json,
-        stats_publisher,
+        args.brokers, args.topic, args.simulation_mode, init_hist_json, stats_publisher
     )
     main.run()
